requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
requests>=2.31

torch>=2.1
torch-geometric>=2.4

rdkit-pypi>=2022.9.5
